refactor (structure) 구조 변경
