+++ src/main/java/nextstep/subway/common/domain/BaseEntity.java
