--- src/main/java/nextstep/subway/exception/CannotFindStationException.java
@@ -1,8 +1,0 @@
-package nextstep.subway.exception;
-
-public class CannotFindStationException extends RuntimeException {
-
-    public CannotFindStationException(String message) {
-        super(message);
-    }
-}
