+++ src/main/java/nextstep/subway/line/dto/LineDto.java
@@ -1,8 +1,9 @@
-package nextstep.subway.dto;
+package nextstep.subway.line.dto;
 
-import nextstep.subway.domain.Line;
-import nextstep.subway.domain.LineBridge;
-import nextstep.subway.domain.Station;
+import nextstep.subway.line.domain.Line;
+import nextstep.subway.station.dto.StationResponse;
+import nextstep.subway.linebridge.domain.LineBridge;
+import nextstep.subway.station.domain.Station;
 
 import java.util.HashSet;
 import java.util.List;
