+++ src/main/java/nextstep/subway/station/application/StationService.java
@@ -1,10 +1,9 @@
-package nextstep.subway.application;
+package nextstep.subway.station.application;
 
-import nextstep.subway.domain.Station;
-import nextstep.subway.domain.StationRepository;
-import nextstep.subway.dto.StationRequest;
-import nextstep.subway.dto.StationResponse;
-import nextstep.subway.exception.CannotFindStationException;
+import nextstep.subway.station.domain.Station;
+import nextstep.subway.station.dto.StationRequest;
+import nextstep.subway.station.dto.StationResponse;
+import nextstep.subway.station.repository.StationRepository;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
