+++ src/main/java/nextstep/subway/station/repository/StationRepository.java
@@ -1,8 +1,10 @@
-package nextstep.subway.domain;
+package nextstep.subway.station.repository;
 
+import nextstep.subway.station.domain.Station;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
+
 
 public interface StationRepository extends JpaRepository<Station, Long> {
     @Override
